Turned AbstractComponentRepainter into a super class instead of an interface to just have the methods inheried to it but with a super class a lot of duplicate code was reduced to only execute the logic once in one class instead of two
